feat: crude infinite terrain generation

+++ rust-vox/src/engine/terrain/mod.rs
@@ -1,4 +1,4 @@
-use noise::{Perlin, NoiseFn, Seedable};
+use noise::{Perlin, NoiseFn};
 
 use super::voxel::Voxel;
 
@@ -6,7 +6,7 @@ pub trait TerrainGenerator
 {
     /// Determine the type of block that will reside at the specified x,y,z in the world \
     /// The x,y,z coordinates must be in world coordinates
-    fn generate( &self, voxel: &mut Voxel , x:u32, y:u32, z:u32);
+    fn generate( &self, voxel: &mut Voxel , x:i32, y:i32, z:i32);
 }
 
 pub struct PerlinGenerator
@@ -25,14 +25,14 @@ impl PerlinGenerator
 
 impl TerrainGenerator for PerlinGenerator
 {
-    fn generate( &self, voxel: &mut Voxel,  x:u32, y:u32, z:u32)
+    fn generate( &self, voxel: &mut Voxel,  x:i32, y:i32, z:i32)
     {
         // println!("points we got, x:{} y:{} ", x as f64 * 10.0 , z as f64 * 10.0 );
         let max_height = self.perlin.get([x as f64 / 10.0, z as f64 / 10.0]) * 10.0 + 10.0 ;
-        let max_height = max_height as u32;
+        let max_height = max_height as i32;
         // println!("max height {}" , max_height);
 
-        if y >= max_height as u32
+        if y >= max_height
         {
             voxel.set_filled(false);
             return;
